requirements.txt
flask
librosa
numpy
pandas
tensorflow
matplotlib
SpeechRecognition
PyAudio
openai-whisper